+++ tests/Unit/Schema/ResolverProviderTest.php
@@ -9,6 +9,7 @@ use Nuwave\Lighthouse\Schema\Values\TypeValue;
 use Nuwave\Lighthouse\Schema\AST\PartialParser;
 use Nuwave\Lighthouse\Schema\Values\FieldValue;
 use Nuwave\Lighthouse\Exceptions\DefinitionException;
+use Tests\Utils\Queries\FooBar;
 
 class ResolverProviderTest extends TestCase
 {
@@ -47,6 +48,39 @@ class ResolverProviderTest extends TestCase
         $this->assertInstanceOf(
             Closure::class,
             $this->resolverProvider->provideResolver($fieldValue)
+        );
+    }
+
+    /**
+     * @test
+     */
+    public function itGetsTheConventionBasedDefaultResolverForRootFieldsWithInvoke(): void
+    {
+        $fieldValue = $this->constructFieldValue('fooInvoke: Int');
+
+        $this->assertInstanceOf(
+            Closure::class,
+            $this->resolverProvider->provideResolver($fieldValue)
+        );
+    }
+
+    /**
+     * @test
+     * @deprecated will be changed in v5
+     */
+    public function itGetsTheConventionBasedDefaultResolverForRootFieldsAndDefaultsToResolve(): void
+    {
+        $fieldValue = $this->constructFieldValue('fooBar: String');
+
+        $resolver = $this->resolverProvider->provideResolver($fieldValue);
+        $this->assertInstanceOf(
+            Closure::class,
+            $resolver
+        );
+
+        $this->assertSame(
+            FooBar::RESOLVE_RESULT,
+            $resolver()
         );
     }