+++ tests/Utils/Queries/FooInvoke.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace Tests\Utils\Queries;
+
+class FooInvoke
+{
+    public function __invoke(): int {}
+}